add simd version

+++ src/minimizerseeding.h
@@ -7,19 +7,14 @@
 
 #include "seeding.h"
 #include "util.h"
-#include <cstdint>
-#include <inttypes.h>
-#include <algorithm>
-#include <vector>
-#include <cmath>
-#include <list>
+#include <bits/stdc++.h>
 
 #ifndef _MINIMIZERSEEDING_H
 #define _MINIMIZERSEEDING_H
 
 struct MMentry{
     uint64_t hash;
-    size_t pos;
+    int pos;
     kmer str;
 };
 
@@ -30,18 +25,17 @@ private:
 
     inline uint64_t hash64(kmer s, uint64_t mask);
     int w;
-	
+    
 public:
 
     minimizerseeding(int k1, int w1): seeding(k1 + w1 - 1, k1)
-	{
-	    w = w1;
-	}
+    {
+        w = w1;
+    }
     void get_minimizers(std::string s, std::vector<seed>& seeds);
     //produce seeds for both s rand revComp(s)
     //stored as 0-s_idx.mmseed and 1-s_idx.mmseed respectively
     double getSeeds(std::string& s, const size_t s_idx,
-		    const char* ouput_dir, const int dir_len);
-
+            const char* ouput_dir, const int dir_len);
 };
 #endif
