+++ src/syncmerseeding.h
@@ -7,20 +7,15 @@
 
 #include "seeding.h"
 #include "util.h"
-#include <cstdint>
-#include <inttypes.h>
-#include <algorithm>
-#include <vector>
-#include <cmath>
-#include <list>
+#include <bits/stdc++.h>
 
 #ifndef _SYNCMERSEEDING_H
 #define _SYNCMERSEEDING_H
 
 struct MMentry{
-    uint64_t hash;
-    size_t pos;
-    kmer str;
+	uint64_t hash;
+	int pos;
+	kmer str;
 };
 
 class syncmerseeding: private seeding
@@ -28,21 +23,21 @@ class syncmerseeding: private seeding
 
 private:
 
-    uint64_t murmur64(kmer key);
-    int s;
-    std::vector<int> pos;
+	uint64_t murmur64(kmer key);
+	int s;
+	std::vector<int> pos;
 public:
 
-    syncmerseeding(int k1, int s1): seeding(k1, k1)
+	syncmerseeding(int k1, int s1): seeding(k1, k1)
 	{
-	    s = s1;
+		s = s1;
 	}
-    void add(int p)
+	void add(int p)
 	{
-	    pos.push_back(p);
+		pos.push_back(p);
 	}
 
-    void get_syncmers(std::string s, std::vector<seed>& seeds);
+	void get_syncmers(std::string s, std::vector<seed>& seeds);
     //produce seeds for both s and revComp(s),
     //stored as 0-s_idx.syncmerseed and 1-s_idx.syncmerseed respectively
     double getSeeds(std::string& s, const size_t s_idx,
